fix bug

+++ ccm_backend/src/main/java/nhom12/uth/ccm/repository/IPaymentTransactionRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface IPaymentTransactionRepository extends JpaRepository<PaymentTransaction, Integer> {
+public interface IPaymentTransactionRepository extends JpaRepository<PaymentTransaction, Long> {
 
     List<PaymentTransaction> findByeWalletOrderByCreatedAtDesc(EWallet eWallet);
